Refactored Preset Logic into PresetProvider.tsx
